Cleaned up app.js
