+++ public/stylesheets/css/main.css
@@ -341,38 +341,57 @@ nav {
     margin: auto; }
   .floating-box__large {
     width: 60%;
-    margin: auto; }
-  .floating-box__send-text__content {
-    text-align: center; }
-    .floating-box__send-text__content p {
-      font-size: 13px; }
-    .floating-box__send-text__content__icon {
-      height: 70px;
-      margin-top: 40px; }
-      .floating-box__send-text__content__icon img {
+    margin: auto;
+    padding: 30px 0px;
+    box-sizing: border-box; }
+    .floating-box__large form input, .floating-box__large form textarea {
+      margin: 10px auto 0px auto; }
+    .floating-box__large form input[type="submit"] {
+      margin-top: 30px;
+      display: inline-block;
+      width: 30%; }
+    .floating-box__large__icon {
+      margin: 20px auto; }
+      .floating-box__large__icon img {
         width: 60px;
         height: 60px; }
-    .floating-box__send-text__content form textarea, .floating-box__send-text__content form input {
-      display: block;
-      margin: auto; }
-    .floating-box__send-text__content form textarea {
+  .floating-box__send-text__content, .floating-box__event__content {
+    text-align: center; }
+    .floating-box__send-text__content p, .floating-box__event__content p {
+      font-size: 13px; }
+    .floating-box__send-text__content form textarea, .floating-box__send-text__content form input, .floating-box__event__content form textarea, .floating-box__event__content form input {
+      display: block; }
+    .floating-box__send-text__content form textarea, .floating-box__event__content form textarea {
       margin: 30px auto;
       min-height: 200px;
       width: 80%;
       max-width: 80%;
       max-height: 300px; }
-    .floating-box__send-text__content form input[type="checkbox"], .floating-box__send-text__content form p {
+    .floating-box__send-text__content form input[type="checkbox"], .floating-box__send-text__content form p, .floating-box__event__content form input[type="checkbox"], .floating-box__event__content form p {
       display: inline-block;
       font-size: 16px;
       vertical-align: middle; }
-    .floating-box__send-text__content form input[type="checkbox"] {
+    .floating-box__send-text__content form input[type="checkbox"], .floating-box__event__content form input[type="checkbox"] {
       margin-right: 10px; }
-    .floating-box__send-text__content form .button {
+    .floating-box__send-text__content form .button, .floating-box__event__content form .button {
       margin: 30px auto; }
-    .floating-box__send-text__content form .confirm-text.red {
+    .floating-box__send-text__content form .confirm-text.red, .floating-box__event__content form .confirm-text.red {
       color: red; }
-    .floating-box__send-text__content p:nth-last-child(1) {
+    .floating-box__send-text__content p:nth-last-child(1), .floating-box__event__content p:nth-last-child(1) {
       padding: 0px 10% 35px 10%; }
+  .floating-box__event form {
+    width: 80%;
+    margin: auto; }
+    .floating-box__event form input.input-half:nth-last-child(2) {
+      float: right; }
+    .floating-box__event form input.input-half:nth-last-child(3) {
+      float: left; }
+  .floating-box__event input {
+    width: 100%; }
+  .floating-box__event input.input-half {
+    width: 48%;
+    display: inline-block;
+    vertical-align: top; }
   .floating-box__content {
     text-align: center;
     box-sizing: border-box;
