+++ views/pages/admin-dashboard.ejs
@@ -23,7 +23,7 @@
 
         <section id="admin-controls">
             <div class="admin-controls">
-                <a href="#">
+                <a href="/admin/event">
                     <div class="floating-box admin-controls__item">
                         <img src="/img/icons/add-event.svg"/>
                         <h3>Add event</h3>
