+++ views/pages/create-event.ejs
@@ -16,20 +16,17 @@
 
     <section id="main">
 
-        <div class="floating-box floating-box__large floating-box__send-text vertC__relative">
-            <div class="floating-box__send-text__content">
+        <div class="floating-box floating-box__large floating-box__event vertC__relative">
+            <div class="floating-box__event__content">
                 <div class="floating-box__large__icon">
-                    <img src="/img/icons/send-text.svg"/>
-                    <h3>Send Text</h3>
+                    <img src="/img/icons/add-event.svg"/>
+                    <h3>Create Event</h3>
                 </div>
                 <form id="send-text" method="post">
-                    <textarea name="text" placeholder="Type your message here..."></textarea>
-                    <input id="confirm" type="checkbox" name="confirm" value="yes">
-                    <p class="confirm-text">I confirm that I have checked the above message.</p><br>
-                    <input class="button __blue" type="submit" placeholder="Login"/>
-                    <p>The following message will be sent to all hackathon participants who have provided mobile numbers
-                        on sign up.
-                    </p>
+                    <input type="text" placeholder="Event Name"/>
+                    <input class="input-half" type="text" placeholder="Event Time"/>
+                    <input class="input-half" type="text" placeholder="Where"/>
+                    <input class="button __blue" type="submit" placeholder="ADD"/>
                 </form>
             </div>
         </div>
